refactor: test files to make them simpler
